If the job's output file exists, then write into it; otherwise copy the temporary output file to the job's output file.

+++ src/main/python/conveyor/recipe.py
@@ -565,7 +565,13 @@ class _UnifiedRecipe(_Recipe):
         self._log.info(
             'job %d: copying output: %s -> %s', self._job.id, output_file,
             self._job.output_file)
-        shutil.copy(output_file, self._job.output_file)
+        if not os.path.exists(self._job.output_file):
+            shutil.copy(output_file, self._job.output_file)
+        else:
+            with open(output_file) as read_fp:
+                with open(self._job.output_file, 'w') as write_fp:
+                    for line in read_fp:
+                        write_fp.write(line)
         task.end(None)
